update documentation and dependencies
