+++ input.json
@@ -1,0 +1,64 @@
+{
+    "name": "test",
+    "esdl_type" : "EConnection",
+    "description" : "this is a test description",
+    "relevant_links" : [
+        {
+            "name" : "test link",
+            "url" : "https://example.com/test",
+            "description" : "this is a test link"
+        },
+        {
+            "name" : "another test link",
+            "url" : "https://example.com/anothertest",
+            "description" : "this is another link"
+        }
+    ],
+    "calculations" : [
+        {
+            "name": "test calculation",
+            "description" : "test",
+            "time_period_in_seconds" : 900,
+            "offset_in_seconds" : 0,
+            "inputs" : [
+                {
+                    "name" : "input1",
+                    "esdl_type" : "PVInstallation",
+                    "data_type" : "DOUBLE",
+                    "description" : "input 1 description",
+                    "unit" : "K"
+                }
+            ],
+            "outputs" : [
+                {
+                    "name" : "output1",
+                    "data_type" : "DOUBLE",
+                    "description" : "output 1 description",
+                    "unit" : "W"
+                },
+                {
+                    "name" : "output2",
+                    "data_type" : "DOUBLE",
+                    "description" : "output 2 description",
+                    "unit" : "W"
+                }
+            ]
+        },
+        {
+            "name": "test calculation 2",
+            "description" : "test",
+            "time_period_in_seconds" : 900,
+            "offset_in_seconds" : 100,
+            "inputs" : [
+            ],
+            "outputs" : [
+                {
+                    "name" : "output3",
+                    "data_type" : "DOUBLE",
+                    "description" : "output 3 description",
+                    "unit" : "W"
+                }
+            ]
+        }
+    ]
+}
