+++ pyproject.toml
@@ -16,6 +16,10 @@ classifiers = [
     "License :: OSI Approved :: MIT License",
     "Operating System :: OS Independent",
 ]
+dependencies = [
+  'helics==3.6.1',
+  'dots_infrastructure==0.4.8'
+]
 
 [project.urls]
 Homepage = "https://github.com/EES-TUe/Dots-calculation-service-template"
